+++ Examples/YandexMobileAdsExample/YandexMobileAdsExample/MobileMediation/MobileMediationInterstitialViewController.swift
@@ -8,27 +8,27 @@
 import UIKit
 import YandexMobileAds
 
-private let adMobBlockID = "adf-279013/975869"
-private let appLovinBlockID = "adf-279013/1052107"
-private let facebookBlockID = "adf-279013/975872"
-private let ironSourceBlockID = "adf-279013/1052109"
-private let moPubBlockID = "adf-279013/975870"
-private let myTargetBlockID = "adf-279013/975871";
-private let startAppBlockID = "adf-279013/1006406"
-private let unityAdsBlockID = "adf-279013/1006439"
-private let yandexBlockID = "adf-279013/975873"
+private let adMobAdUnitID = "adf-279013/975869"
+private let appLovinAdUnitID = "adf-279013/1052107"
+private let facebookAdUnitID = "adf-279013/975872"
+private let ironSourceAdUnitID = "adf-279013/1052109"
+private let moPubAdUnitID = "adf-279013/975870"
+private let myTargetAdUnitID = "adf-279013/975871";
+private let startAppAdUnitID = "adf-279013/1006406"
+private let unityAdsAdUnitID = "adf-279013/1006439"
+private let yandexAdUnitID = "adf-279013/975873"
 
 class MobileMediationInterstitialViewController: UIViewController {
-    private let blockIDs = [
-        (adapter: "AdMob", blockID: adMobBlockID),
-        (adapter: "AppLovin", blockID: appLovinBlockID),
-        (adapter: "Facebook", blockID: facebookBlockID),
-        (adapter: "IronSource", blockID: ironSourceBlockID),
-        (adapter: "MoPub", blockID: moPubBlockID),
-        (adapter: "myTarget", blockID: myTargetBlockID),
-        (adapter: "StartApp", blockID: startAppBlockID),
-        (adapter: "UnityAds", blockID: unityAdsBlockID),
-        (adapter: "Yandex", blockID: yandexBlockID)
+    private let adUnitIDs = [
+        (adapter: "AdMob", adUnitID: adMobAdUnitID),
+        (adapter: "AppLovin", adUnitID: appLovinAdUnitID),
+        (adapter: "Facebook", adUnitID: facebookAdUnitID),
+        (adapter: "IronSource", adUnitID: ironSourceAdUnitID),
+        (adapter: "MoPub", adUnitID: moPubAdUnitID),
+        (adapter: "myTarget", adUnitID: myTargetAdUnitID),
+        (adapter: "StartApp", adUnitID: startAppAdUnitID),
+        (adapter: "UnityAds", adUnitID: unityAdsAdUnitID),
+        (adapter: "Yandex", adUnitID: yandexAdUnitID)
     ]
     
     @IBOutlet private var showButton: UIButton!
@@ -44,20 +44,20 @@ class MobileMediationInterstitialViewController: UIViewController {
         self.showButton.isEnabled = false
         let selectedIndex = pickerView.selectedRow(inComponent: 0)
         /*
-         Replace demo AdMobBlockID with actual Block ID.
-         Following demo block ids may be used for testing:
-         AdMob mediation: adMobBlockID
-         AppLovin mediation: appLovinBlockID
-         Facebook mediation: facebookBlockID
-         IronSource mediation: ironSourceBlockID
-         MoPub mediation: moPubBlockID
-         MyTarget mediation: myTargetBlockID
-         StartApp mediation: startAppBlockID
-         UnityAds mediation: unityAdsBlockID
-         Yandex: yandexBlockID
+         Replace demo adUnitID with actual Ad Unit ID.
+         Following demo ad unit ids may be used for testing:
+         AdMob mediation: adMobAdUnitID
+         AppLovin mediation: appLovinAdUnitID
+         Facebook mediation: facebookAdUnitID
+         IronSource mediation: ironSourceAdUnitID
+         MoPub mediation: moPubAdUnitID
+         MyTarget mediation: myTargetAdUnitID
+         StartApp mediation: startAppAdUnitID
+         UnityAds mediation: unityAdsAdUnitID
+         Yandex: yandexAdUnitID
          */
-        let blockID = blockIDs[selectedIndex].blockID
-        interstitialAd = YMAInterstitialAd(blockID: blockID)
+        let adUnitID = adUnitIDs[selectedIndex].adUnitID
+        interstitialAd = YMAInterstitialAd(adUnitID: adUnitID)
         interstitialAd?.delegate = self
         interstitialAd?.load()
     }
@@ -116,7 +116,7 @@ extension MobileMediationInterstitialViewController: YMAInterstitialAdDelegate {
 
 extension MobileMediationInterstitialViewController: UIPickerViewDelegate, UIPickerViewDataSource {
     func pickerView(_ pickerView: UIPickerView, numberOfRowsInComponent component: Int) -> Int {
-        return blockIDs.count
+        return adUnitIDs.count
     }
 
     func numberOfComponents(in pickerView: UIPickerView) -> Int {
@@ -124,6 +124,6 @@ extension MobileMediationInterstitialViewController: UIPickerViewDelegate, UIPic
     }
 
     func pickerView(_ pickerView: UIPickerView, titleForRow row: Int, forComponent component: Int) -> String? {
-        return blockIDs[row].adapter
+        return adUnitIDs[row].adapter
     }
 }