+++ changelog/adapter/admob-mobileads/CHANGELOG.md
@@ -1,6 +1,13 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## Version 5.0.0.0
+
+#### Updated
+* Added support for Yandex Mobile Ads SDK 5.0.0
+* Added support for Google Mobile Ads SDK 9.0.0
+* Updated minimum supported Google Mobile Ads SDK version to 9.0.0
+
 ## Version 4.4.0.0
 
 #### Updated